+++ apps/fabric/metro.config.js
@@ -9,6 +9,8 @@ const root = path.resolve(__dirname, '../../packages/core/');
 
 const modules = Object.keys(pack.peerDependencies);
 
+const defaultConfig = getDefaultConfig(__dirname);
+
 /**
  * Metro configuration
  * https://facebook.github.io/metro/docs/configuration
@@ -22,6 +24,7 @@ const config = {
   // We need to make sure that only one version is loaded for peerDependencies
   // So we exclude them at the root, and alias them to the versions in example's node_modules
   resolver: {
+    assetExts: [...defaultConfig.resolver.assetExts, 'lottie'],
     blacklistRE: exclusionList(
       modules.map(
         m => new RegExp(`^${escape(path.join(root, 'node_modules', m))}\\/.*$`),
@@ -35,4 +38,4 @@ const config = {
   },
 };
 
-module.exports = mergeConfig(getDefaultConfig(__dirname), config);
+module.exports = mergeConfig(defaultConfig, config);