+++ packages/core/ios/Fabric/LottieContainerView.h
@@ -23,6 +23,7 @@
 - (void)setSourceURL:(NSString * _Nonnull)newSourceURLString;
 - (void)setSourceJson:(NSString * _Nonnull)newSourceJson;
 - (void)setSourceName:(NSString * _Nonnull)newSourceName;
+- (void)setSourceDotLottie:(NSDictionary * _Nonnull)source;
 - (void)setResizeMode:(NSString * _Nonnull)resizeMode;
 - (void)setColorFilters:(NSArray<NSDictionary *> * _Nonnull)newColorFilters;
 - (void)play;